Try to render a PDF page
I was seeing requests for files that we’re not serving. They’re
in the /build directory, though. So we want the bundle to be
asking for /static/foo.js when it’s actually asking for /foo.js.
I’m not sure whether that’s a problem with the bundle, the
server, or something else. — SK